+++ selfdrive/ui/SConscript
@@ -58,7 +58,7 @@ qt_env.Program("qt/spinner", ["qt/spinner.cc"], LIBS=qt_libs)
 # build main UI
 qt_src = ["main.cc", "ui.cc", "qt/sidebar.cc", "qt/onroad.cc",
           "qt/window.cc", "qt/home.cc", "qt/offroad/settings.cc",
-          "qt/offroad/onboarding.cc", "qt/offroad/driverview.cc", "qt/wakeable.cc"]
+          "qt/offroad/onboarding.cc", "qt/offroad/driverview.cc"]
 qt_env.Program("_ui", qt_src + [asset_obj], LIBS=qt_libs)
